Revert "Added testing with Jest"
