+++ index.html
@@ -18,13 +18,12 @@
     <script type="text/javascript" src="./lib/ml5.min.js"></script>
     <script type="text/javascript" src="./lib/Tone.js"></script>
 
-    <script type="text/javascript" src="velato_constants.js"></script>
-    <script type="text/javascript" src="velato_token.js"></script>
-    <script type="text/javascript" src="velato_note.js"></script>
-    <script type="text/javascript" src="velato_web_display.js"></script>
-    <script type="text/javascript" src="velato_programbuilder.js"></script>
-    <script type="text/javascript" src="velato_audio_interface.js"></script>
-    <script type="text/javascript" src="page_controls.js"></script>
+    <script type="text/javascript" type="module" src="velato_constants.js"></script>
+    <script type="text/javascript" type="module" src="velato_token.js"></script>
+    <script type="text/javascript" type="module" src="velato_note.js"></script>
+    <script type="text/javascript" type="module" src="velato_programbuilder.js"></script>
+    <script type="text/javascript" type="module" src="velato_audio_interface.js"></script>
+    <script type="text/javascript" type="module" src="page_controls.js"></script>
 </head>
 <body>
 <div id="outframe">
