--- package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "jest": "^30.0.2"
-  }
-}
